requirements.txt
pytest
mysql-connector-python
pre-commit